+++ packages/ui/spa/components/designSystem/avatar.tsx
